Rename HTMLParser to Parser
